# docs/swagger.yaml
basePath: /
definitions:
...
  levelConfig.LevelInfo:
    properties:
      isMaster:
        type: boolean
      levelName:
        type: string
      leveldataoverride:
        type: string
      modoverrides:
        type: string
      runVersion:
        type: integer
      server_ini:
        $ref: '#/definitions/levelConfig.ServerIni'
      uuid:
        type: string
    type: object
  levelConfig.ServerIni:
    properties:
      authentication_port:
        description: '[STEAM]'
        type: integer
      encode_user_path:
        description: '[ACCOUNT]'
        type: boolean
      id:
        type: integer
      is_master:
        description: '[SHARD]'
        type: boolean
      master_server_port:
        type: integer
      name:
        type: string
      server_port:
        description: '[NETWORK]'
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        description: 提示代码
        type: integer
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
host: localhost:8082
info:
  contact:
    name: API Support
    url: https://github.com/carrot-hu23/dst-admin-go
  description: 饥荒联机版服务器管理后台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DST Admin Go API
  version: "1.7.0"
paths:
  /api/cluster/level:
    delete:
      consumes:
      - application/json
      description: 删除指定的世界
      parameters:
      - description: 世界名称
        in: query
        name: levelName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除世界
      tags:
      - level
    get:
      consumes:
      - application/json
      description: 获取当前集群的所有世界(等级)列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/levelConfig.LevelInfo'
                  type: array
              type: object
      summary: 获取世界列表
      tags:
      - level
    post:
      consumes:
      - application/json
      description: 创建一个新的世界(等级)
      parameters:
      - description: 世界配置信息
        in: body
        name: level
        required: true
        schema:
          $ref: '#/definitions/levelConfig.LevelInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建世界
      tags:
      - level
    put:
      consumes:
      - application/json
      description: 批量更新多个世界的配置信息
      parameters:
      - description: 世界配置信息列表
        in: body
        name: levels
        required: true
        schema:
          items:
            $ref: '#/definitions/levelConfig.LevelInfo'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 批量更新世界
      tags:
      - level
...
  /api/kv:
    get:
      description: 获取kv值
      parameters:
      - description: key
        in: query
        name: key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取kv值
      tags:
      - kv
    post:
      description: 保存kv值
      parameters:
      - description: key
        in: formData
        name: key
        required: true
        type: string
      - description: value
        in: formData
        name: value
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 保存kv值
      tags:
      - kv
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"